app/schemas/User.py
from pydantic import BaseModel

class UserBase(BaseModel):
    username: str


class UserCreate(UserBase):
    password: str


class UserUpdate(BaseModel):
    username: str | None = None
    password: str | None = None


class UserResponse(UserBase):
    id: int
    
    model_config = {
        "from_attributes": True
    }